STM8: logger warnings.

+++ tests/kern2.c
@@ -83,102 +83,102 @@ uint32_t test_start (void)
     /* Check all variables contain expected values */
     if (one != 1)
     {
-        ATOMLOG (_STR("1(%d)\n"), one);
+        ATOMLOG (_STR("1(%d)\n"), (int)one);
         failures++;
     }
     if (two != 2)
     {
-        ATOMLOG (_STR("2(%d)\n"), two);
+        ATOMLOG (_STR("2(%d)\n"), (int)two);
         failures++;
     }
     if (three != 3)
     {
-        ATOMLOG (_STR("3(%d)\n"), three);
+        ATOMLOG (_STR("3(%d)\n"), (int)three);
         failures++;
     }
     if (four != 4)
     {
-        ATOMLOG (_STR("4(%d)\n"), four);
+        ATOMLOG (_STR("4(%d)\n"), (int)four);
         failures++;
     }
     if (five != 5)
     {
-        ATOMLOG (_STR("5(%d)\n"), five);
+        ATOMLOG (_STR("5(%d)\n"), (int)five);
         failures++;
     }
     if (six != 6)
     {
-        ATOMLOG (_STR("6(%d)\n"), six);
+        ATOMLOG (_STR("6(%d)\n"), (int)six);
         failures++;
     }
     if (seven != 7)
     {
-        ATOMLOG (_STR("7(%d)\n"), seven);
+        ATOMLOG (_STR("7(%d)\n"), (int)seven);
         failures++;
     }
     if (eight != 8)
     {
-        ATOMLOG (_STR("8(%d)\n"), eight);
+        ATOMLOG (_STR("8(%d)\n"), (int)eight);
         failures++;
     }
     if (nine != 9)
     {
-        ATOMLOG (_STR("9(%d)\n"), nine);
+        ATOMLOG (_STR("9(%d)\n"), (int)nine);
         failures++;
     }
     if (ten != 10)
     {
-        ATOMLOG (_STR("10(%d)\n"), ten);
+        ATOMLOG (_STR("10(%d)\n"), (int)ten);
         failures++;
     }
     if (eleven != 11)
     {
-        ATOMLOG (_STR("11(%d)\n"), eleven);
+        ATOMLOG (_STR("11(%d)\n"), (int)eleven);
         failures++;
     }
     if (twelve != 12)
     {
-        ATOMLOG (_STR("12(%d)\n"), twelve);
+        ATOMLOG (_STR("12(%d)\n"), (int)twelve);
         failures++;
     }
     if (thirteen != 13)
     {
-        ATOMLOG (_STR("13(%d)\n"), thirteen);
+        ATOMLOG (_STR("13(%d)\n"), (int)thirteen);
         failures++;
     }
     if (fourteen != 14)
     {
-        ATOMLOG (_STR("14(%d)\n"), fourteen);
+        ATOMLOG (_STR("14(%d)\n"), (int)fourteen);
         failures++;
     }
     if (fifteen != 15)
     {
-        ATOMLOG (_STR("15(%d)\n"), fifteen);
+        ATOMLOG (_STR("15(%d)\n"), (int)fifteen);
         failures++;
     }
     if (sixteen != 16)
     {
-        ATOMLOG (_STR("16(%d)\n"), sixteen);
+        ATOMLOG (_STR("16(%d)\n"), (int)sixteen);
         failures++;
     }
     if (seventeen != 17)
     {
-        ATOMLOG (_STR("17(%d)\n"), seventeen);
+        ATOMLOG (_STR("17(%d)\n"), (int)seventeen);
         failures++;
     }
     if (eighteen != 18)
     {
-        ATOMLOG (_STR("18(%d)\n"), eighteen);
+        ATOMLOG (_STR("18(%d)\n"), (int)eighteen);
         failures++;
     }
     if (nineteen != 19)
     {
-        ATOMLOG (_STR("19(%d)\n"), nineteen);
+        ATOMLOG (_STR("19(%d)\n"), (int)nineteen);
         failures++;
     }
     if (twenty != 20)
     {
-        ATOMLOG (_STR("20(%d)\n"), twenty);
+        ATOMLOG (_STR("20(%d)\n"), (int)twenty);
         failures++;
     }
